+++ client-v2/src/utils.ts
@@ -1,0 +1,6 @@
+export const toCustomDateStr = (d: string): string => {
+  return new Date(d).toLocaleDateString(undefined, {
+    month: "short",
+    day: "2-digit",
+  });
+};
